Example1: Cleaned up the parser code and prepared for example 1.
