+++ web/common/common.dart
@@ -8,4 +8,9 @@ import 'dart:convert' as convert;
 
 import 'package:ThreeDart/Tokenizer.dart' as Tokenizer;
 
+part 'codeParser.dart';
+part 'dartParser.dart';
+part 'glslParser.dart';
+part 'htmlParser.dart';
+part 'plainParser.dart';
 part 'shellPage.dart';
