+++ web/common/shellPage.dart
@@ -3,13 +3,8 @@ part of ThreeDart.web.common;
 /// The shell page is a tool for creating pages quickly and
 /// easily which will have a consistent look and feel.
 class ShellPage {
-
-  convert.HtmlEscape _escape;
   html.DivElement _page;
   Tokenizer.Tokenizer _parTokenizer;
-  Tokenizer.Tokenizer _htmlTokenizer;
-  Tokenizer.Tokenizer _dartTokenizer;
-  Tokenizer.Tokenizer _glslTokenizer;
 
   /// Creates a new shell page with an optional [title].
   ShellPage([String title = ""]) {
@@ -35,12 +30,9 @@ class ShellPage {
       pageCenter.append(titleElem);
     }
 
-    this._escape = new convert.HtmlEscape(convert.HtmlEscapeMode.ELEMENT);
     this._page = new html.DivElement();
     pageCenter.append(this._page);
-    this._htmlTokenizer = null;
-    this._dartTokenizer = null;
-    this._glslTokenizer = null;
+    this._parTokenizer = null;
 
     html.document.onScroll.listen((_) {
     	scrollTop.style.top = "${-0.05*body.scrollTop}px";
@@ -127,6 +119,20 @@ class ShellPage {
     this._page.append(parElem);
   }
 
+  /// Gets the code parser for the given [lang].
+  CodeParser getCodeParser(String lang) {
+    // Full list of parsers, add more as needed.
+    List<CodeParser> parsers = [
+      new DartParser(),
+      new GLSLParser(),
+      new HTMLParser()
+    ];
+
+    CodeParser parser = parsers.firstWhere((CodeParser parser) => parser.name == lang);
+    if (parser != null) return parser;
+    return new PlainParser();
+  }
+
   /// Adds a code box with the given [title] to the page.
   /// The given [lang] is the language to color the code with.
   /// Currently it supports HTML, Dart, GLSL, and other.
@@ -149,12 +155,8 @@ class ShellPage {
       } else diff.add(0);
     }
 
-    List<List<html.DivElement>> lineList = [];
-    String code = lines.join("\n");
-    if (lang == "html")      this._colorHtml(code, lineList);
-    else if (lang == "dart") this._colorDart(code, lineList);
-    else if (lang == "glsl") this._colorGlsl(code, lineList);
-    else                     this._colorPlain(code, lineList);
+    CodeParser colorCode = this.getCodeParser(lang);
+    colorCode.parse(lines);
 
     html.DivElement codeTableScroll = new html.DivElement()
       ..className = "codeTableScroll";
@@ -182,8 +184,8 @@ class ShellPage {
 
     if (showDiff) {
       int lineNoSub = firstLineNo, lineNoAdd = firstLineNo;
-      for (int i = 0; i < lineList.length; ++i) {
-        List<html.DivElement> line = lineList[i];
+      for (int i = 0; i < colorCode.lineList.length; ++i) {
+        List<html.DivElement> line = colorCode.lineList[i];
         html.TableRowElement rowElem = new html.TableRowElement()
           ..className = "codeTableRow";
         html.TableCellElement cell1Elem = new html.TableCellElement()
@@ -225,7 +227,7 @@ class ShellPage {
       }
     } else {
       int lineNo = firstLineNo;
-      for (List<html.DivElement> line in lineList) {
+      for (List<html.DivElement> line in colorCode.lineList) {
         html.TableRowElement rowElem = new html.TableRowElement()
           ..className = "codeTableRow";
         html.TableCellElement cell1Elem = new html.TableCellElement()
@@ -266,87 +268,6 @@ class ShellPage {
       ..className = "pageCanvas"
       ..id = id;
     this._page.append(canvas);
-  }
-
-  /// Escapes the given [text] for html.
-  String _escapeText(String text) {
-    return this._escape.convert(text).replaceAll(" ", "&nbsp;");
-  }
-
-  /// Adds line parts to the list of code lines, the [lineList].
-  /// The given [code] to add as lines in the given [color].
-  void _addLineParts(String code, String color, List<List<html.DivElement>> lineList) {
-    if (lineList.isEmpty) lineList.add(new List<html.DivElement>());
-    List<String> lines = code.split("\n");
-    bool first = true;
-    for (String line in lines) {
-      if (first) first = false;
-      else lineList.add(new List<html.DivElement>());
-      html.DivElement partElem = new html.DivElement()
-        ..className = "codePart"
-        ..innerHtml = this._escapeText(line)
-        ..style.color = color;
-      lineList.last.add(partElem);
-    }
-  }
-
-  /// Adds color for the given HTML [code] to the given [lineList].
-  void _colorHtml(String code, List<List<html.DivElement>> lineList) {
-    this._setupHtmlTokenizer();
-    for (Tokenizer.Token token in this._htmlTokenizer.tokenize(code)) {
-      switch(token.name) {
-        case "Attr":
-          this._addLineParts(token.text, "#911", lineList);
-          this._addLineParts("=", "#111", lineList);
-          break;
-        case "Id":       this._addLineParts(token.text, "#111", lineList); break;
-        case "Other":    this._addLineParts(token.text, "#111", lineList); break;
-        case "Reserved": this._addLineParts(token.text, "#119", lineList); break;
-        case "String":   this._addLineParts(token.text, "#171", lineList); break;
-        case "Symbol":   this._addLineParts(token.text, "#616", lineList); break;
-      }
-    }
-  }
-
-  /// Adds color for the given Dart [code] to the given [lineList].
-  void _colorDart(String code, List<List<html.DivElement>> lineList) {
-    this._setupDartTokenizer();
-    for (Tokenizer.Token token in this._dartTokenizer.tokenize(code)) {
-      switch(token.name) {
-        case "Class":      this._addLineParts(token.text, "#551", lineList); break;
-        case "Comment":    this._addLineParts(token.text, "#777", lineList); break;
-        case "Id":         this._addLineParts(token.text, "#111", lineList); break;
-        case "Num":        this._addLineParts(token.text, "#191", lineList); break;
-        case "Reserved":   this._addLineParts(token.text, "#119", lineList); break;
-        case "String":     this._addLineParts(token.text, "#171", lineList); break;
-        case "Symbol":     this._addLineParts(token.text, "#616", lineList); break;
-        case "Type":       this._addLineParts(token.text, "#B11", lineList); break;
-        case "Whitespace": this._addLineParts(token.text, "#111", lineList); break;
-      }
-    }
-  }
-
-  /// Adds color for the given GLSL [code] to the given [lineList].
-  void _colorGlsl(String code, List<List<html.DivElement>> lineList) {
-    this._setupGlslTokenizer();
-    for (Tokenizer.Token token in this._glslTokenizer.tokenize(code)) {
-      switch(token.name) {
-        case "Builtin":    this._addLineParts(token.text, "#411", lineList); break;
-        case "Comment":    this._addLineParts(token.text, "#777", lineList); break;
-        case "Id":         this._addLineParts(token.text, "#111", lineList); break;
-        case "Num":        this._addLineParts(token.text, "#191", lineList); break;
-        case "Preprocess": this._addLineParts(token.text, "#737", lineList); break;
-        case "Reserved":   this._addLineParts(token.text, "#119", lineList); break;
-        case "Symbol":     this._addLineParts(token.text, "#611", lineList); break;
-        case "Type":       this._addLineParts(token.text, "#171", lineList); break;
-        case "Whitespace": this._addLineParts(token.text, "#111", lineList); break;
-      }
-    }
-  }
-
-  /// Adds color for the given plain [code] to the given [lineList].
-  void _colorPlain(String code, List<List<html.DivElement>> lineList) {
-    this._addLineParts(code, "#111", lineList);
   }
 
   /// Constructs the paragraph tokenizer if the tokenizer hasn't been setup yet.
@@ -405,237 +326,5 @@ class ShellPage {
     tok.setToken("LinkEnd", "Link");
     tok.setToken("Other", "Other");
     this._parTokenizer = tok;
-  }
-
-  /// Constructs the HTML code tokenizer if the tokenizer hasn't
-  /// been setup yet. The HTML code tokenizer breaks up code to
-  /// label tokens to color the code appropriately.
-  void _setupHtmlTokenizer() {
-    if (this._htmlTokenizer != null) return;
-    Tokenizer.Tokenizer tok = new Tokenizer.Tokenizer();
-    tok.start("Start");
-    tok.join("Start", "Id")
-      ..addSet("_")
-      ..addRange("a", "z")
-      ..addRange("A", "Z");
-    tok.join("Id", "Id")
-      ..addSet("_")
-      ..addRange("0", "9")
-      ..addRange("a", "z")
-      ..addRange("A", "Z");
-    tok.join("Id", "Attr")
-      ..addSet("=")
-      ..consume = true;
-    tok.join("Start", "Sym")
-      ..addSet("</\\-!>=");
-    tok.join("Sym", "Sym")
-      ..addSet("</\\-!>=");
-    tok.join("Start", "OpenStr")
-      ..addSet("\"");
-    tok.join("OpenStr", "CloseStr")
-      ..addSet("\"");
-    tok.join("OpenStr", "EscStr")
-      ..addSet("\\");
-    tok.join("EscStr", "OpenStr")
-      ..addSet("\"");
-    tok.join("OpenStr", "OpenStr")
-      ..addAll();
-    tok.join("Start", "Other")
-      ..addAll();
-    tok.join("Other", "Other").addNot()
-      ..addSet("</\\-!>=_\"")
-      ..addRange("a", "z")
-      ..addRange("A", "Z");
-    tok.setToken("Sym", "Symbol");
-    tok.setToken("CloseStr", "String");
-    tok.setToken("Id", "Id")
-      ..replace("Reserved", ["DOCTYPE", "html", "head", "meta",
-                             "link", "title", "body", "script"]);
-    tok.setToken("Attr", "Attr");
-    tok.setToken("Other", "Other");
-    this._htmlTokenizer = tok;
-  }
-
-  /// Constructs the Dart code tokenizer if the tokenizer hasn't
-  /// been setup yet. The Dart code tokenizer breaks up code to
-  /// label tokens to color the code appropriately.
-  void _setupDartTokenizer() {
-    if (this._dartTokenizer != null) return;
-    Tokenizer.Tokenizer tok = new Tokenizer.Tokenizer();
-    tok.start("Start");
-    tok.join("Start", "Id")
-      ..addSet("_")
-      ..addRange("a", "z")
-      ..addRange("A", "Z");
-    tok.join("Id", "Id")
-      ..addSet("_")
-      ..addRange("0", "9")
-      ..addRange("a", "z")
-      ..addRange("A", "Z");
-    tok.join("Start", "Int")
-      ..addRange("0", "9");
-    tok.join("Int", "Int")
-      ..addRange("0", "9");
-    tok.join("Int", "FloatDot")
-      ..addSet(".");
-    tok.join("FloatDot", "Float")
-      ..addRange("0", "9");
-    tok.join("Float", "Float")
-      ..addRange("0", "9");
-    tok.join("Start", "Sym")
-      ..addSet("<>{}()\\-+*%!&|=.,?:;");
-    tok.join("Sym", "Sym")
-      ..addSet("<>{}()\\-+*%!&|=.,?:;");
-    tok.join("Start", "OpenDoubleStr")
-      ..addSet("\"");
-    tok.join("OpenDoubleStr", "CloseDoubleStr")
-      ..addSet("\"");
-    tok.join("OpenDoubleStr", "EscDoubleStr")
-      ..addSet("\\");
-    tok.join("EscDoubleStr", "OpenDoubleStr")
-      ..addSet("\"");
-    tok.join("OpenDoubleStr", "OpenDoubleStr")
-      ..addAll();
-    tok.join("Start", "OpenSingleStr")
-      ..addSet("'");
-    tok.join("OpenSingleStr", "CloseSingleStr")
-      ..addSet("'");
-    tok.join("OpenSingleStr", "EscSingleStr")
-      ..addSet("\\");
-    tok.join("EscSingleStr", "OpenSingleStr")
-      ..addSet("'");
-    tok.join("OpenSingleStr", "OpenSingleStr")
-      ..addAll();
-    tok.join("Start", "Slash")
-      ..addSet("/");
-    tok.join("Slash", "Comment")
-      ..addSet("/");
-    tok.join("Comment", "EndComment")
-      ..addSet("\n");
-    tok.join("Comment", "Comment")
-      ..addNot().addSet("\n");
-    tok.join("Slash", "MLComment")
-      ..addSet("*");
-    tok.join("MLComment", "MLCStar")
-      ..addSet("*");
-    tok.join("MLCStar", "MLComment")
-      ..addNot().addSet("*");
-    tok.join("MLCStar", "EndComment")
-      ..addSet("/");
-    tok.join("Start", "Whitespace")
-      ..addSet(" \n\t");
-    tok.join("Whitespace", "Whitespace")
-      ..addSet(" \n\t");
-    tok.setToken("Int", "Num");
-    tok.setToken("Float", "Num");
-    tok.setToken("Sym", "Symbol");
-    tok.setToken("CloseDoubleStr", "String");
-    tok.setToken("CloseSingleStr", "String");
-    tok.setToken("EndComment", "Comment");
-    tok.setToken("Whitespace", "Whitespace");
-    tok.setToken("Id", "Id")
-      ..replace("Class", ["Constant", "Depth", "Entity",
-        "EntityPass", "Math", "Matrix4", "Movers", "Rotater",
-        "Scenes", "Shapes", "Techniques", "ThreeDart"])
-      ..replace("Type", ["bool", "double", "dynamic", "false", "int",
-        "List", "Map", "null", "num", "Object", "String", "this",
-        "true", "var", "void"])
-      ..replace("Reserved", ["abstract", "as", "assert", "async", "await",
-        "break", "case", "catch", "class", "continue", "const", "default",
-        "deferred", "do", "else", "enum", "export", "extends", "external",
-        "factory", "final", "finally", "for", "get", "if", "implements",
-        "import", "in", "is", "library", "new", "operator", "part", "rethrow",
-        "return", "set", "static", "super", "switch", "sync", "throw", "try",
-        "typedef", "with", "while", "yield"]);
-    this._dartTokenizer = tok;
-  }
-
-  /// Constructs the GLSL code tokenizer if the tokenizer hasn't
-  /// been setup yet. The GLSL code tokenizer breaks up code to
-  /// label tokens to color the code appropriately.
-  void _setupGlslTokenizer() {
-    if (this._glslTokenizer != null) return;
-    Tokenizer.Tokenizer tok = new Tokenizer.Tokenizer();
-    tok.start("Start");
-    tok.join("Start", "Id")
-      ..addSet("_")
-      ..addRange("a", "z")
-      ..addRange("A", "Z");
-    tok.join("Id", "Id")
-      ..addSet("_")
-      ..addRange("0", "9")
-      ..addRange("a", "z")
-      ..addRange("A", "Z");
-    tok.join("Start", "Int")
-      ..addRange("0", "9");
-    tok.join("Int", "Int")
-      ..addRange("0", "9");
-    tok.join("Int", "FloatDot")
-      ..addSet(".");
-    tok.join("FloatDot", "Float")
-      ..addRange("0", "9");
-    tok.join("Float", "Float")
-      ..addRange("0", "9");
-    tok.join("Start", "Sym")
-      ..addSet("<>{}()\\-+*%!&|=.,?:;");
-    tok.join("Sym", "Sym")
-      ..addSet("<>{}()\\-+*%!&|=.,?:;");
-    tok.join("Start", "Slash")
-      ..addSet("/");
-    tok.join("Slash", "Comment")
-      ..addSet("/");
-    tok.join("Comment", "EndComment")
-      ..addSet("\n");
-    tok.join("Comment", "Comment")
-      ..addNot().addSet("\n");
-    tok.join("Start", "Preprocess")
-      ..addSet("#");
-    tok.join("Preprocess", "Preprocess")
-      ..addNot().addSet("\n");
-    tok.join("Preprocess", "EndPreprocess")
-      ..addSet("\n");
-    tok.join("Start", "Whitespace")
-      ..addSet(" \n\t");
-    tok.join("Whitespace", "Whitespace")
-      ..addSet(" \n\t");
-    tok.setToken("Int", "Num");
-    tok.setToken("Float", "Num");
-    tok.setToken("Sym", "Symbol");
-    tok.setToken("EndComment", "Comment");
-    tok.setToken("EndPreprocess", "Preprocess");
-    tok.setToken("Whitespace", "Whitespace");
-    tok.setToken("Id", "Id")
-      ..replace("Type", [
-        "float", "double", "int", "void", "bool", "true", "false",
-        "mat2", "mat3", "mat4", "dmat2", "dmat3", "dmat4",
-        "mat2x2", "mat2x3", "mat2x4", "dmat2x2", "dmat2x3", "dmat2x4",
-        "mat3x2", "mat3x3", "mat3x4", "dmat3x2", "dmat3x3", "dmat3x4",
-        "mat4x2", "mat4x3", "mat4x4", "dmat4x2", "dmat4x3", "dmat4x4",
-        "vec2", "vec3", "vec4", "ivec2", "ivec3", "ivec4", "bvec2", "bvec3",
-        "bvec4", "dvec2", "dvec3", "dvec4", "uint", "uvec2", "uvec3", "uvec4",
-        "sampler1D", "sampler2D", "sampler3D", "samplerCube",
-        "sampler1DShadow", "sampler2DShadow", "samplerCubeShadow",
-        "sampler1DArray", "sampler2DArray",
-        "sampler1DArrayShadow", "sampler2DArrayShadow",
-        "isampler1D", "isampler2D", "isampler3D", "isamplerCube",
-        "isampler1DArray", "isampler2DArray",
-        "usampler1D", "usampler2D", "usampler3D", "usamplerCube",
-        "usampler1DArray", "usampler2DArray",
-        "sampler2DRect", "sampler2DRectShadow",
-        "isampler2DRect", "usampler2DRect",
-        "samplerBuffer", "isamplerBuffer", "usamplerBuffer",
-        "sampler2DMS", "isampler2DMS", "usampler2DMS",
-        "sampler2DMSArray", "isampler2DMSArray", "usampler2DMSArray",
-        "samplerCubeArray", "samplerCubeArrayShadow",
-        "isamplerCubeArray", "usamplerCubeArray"])
-      ..replace("Reserved", [
-        "attribute", "break", "case", "centroid", "const",
-        "continue", "default", "discard", "do", "else", "flat", "for",
-        "highp", "if", "in", "inout", "invariant", "layout", "lowp",
-        "mediump", "noperspective", "out", "patch", "precision", "return",
-        "sample", "smooth", "struct", "subroutine", "switch", "uniform",
-        "varying", "while"])
-      ..replace("Builtin", ["gl_FragColor", "gl_Position"]);
-    this._glslTokenizer = tok;
   }
 }
