+++ web/examples/example1.dart
@@ -1,0 +1,6 @@
+part of ThreeDart.web.examples;
+
+void example1() {
+  new common.ShellPage("Example 1")
+    ..addPar(["WIP"]);
+}
