+++ web/styles.css
@@ -11,7 +11,7 @@ body {
   left: 0px;
   top: 0px;
   z-index: -1;
-  background-image: url('../resources/SnowTop.png');
+  background-image: url('resources/SnowTop.png');
   background-repeat: no-repeat;
 }
